backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
django-environ==0.11.2
bcrypt==4.1.1
Pillow>=10.2.0
# psycopg2-binary==2.9.9  # Uncomment for PostgreSQL support (requires compatible Python version)

